Fim do algoritmo3. falta descobrir num de cores e fazer o makefile

+++ Ep1/src/ep1/ed/Gemeos.java
@@ -1,22 +1,20 @@
 package ep1.ed;
 
-import java.math.BigInteger;
-
 public class Gemeos {
 
-	private final BigInteger primo1;
-	private final BigInteger primo2;
+	private final long primo1;
+	private final long primo2;
 
-	public Gemeos(BigInteger primo1, BigInteger primo2) {
+	public Gemeos(long primo1, long primo2) {
 		this.primo1 = primo1;
 		this.primo2 = primo2;
 	}
 
-	public BigInteger getPrimo1() {
+	public long getPrimo1() {
 		return primo1;
 	}
-	
-	public BigInteger getPrimo2() {
+
+	public long getPrimo2() {
 		return primo2;
 	}
 }
